Rename: Level -> FilterLevel, LevelUnion -> Level, CategoryUnion -> Category 타입명 수정
