+++ src/constants/type-label-map.ts
@@ -1,6 +1,6 @@
-import { CategoryUnion, LevelUnion } from '@/apis/models/video';
+import { Category, Level } from '@/apis/models/video';
 
-export const categoryItems = new Map<CategoryUnion, string>([
+export const categoryItems = new Map<Category, string>([
   ['FE', '프론트엔드'],
   ['BE', '백엔드'],
   ['AI', 'AI'],
@@ -8,7 +8,7 @@ export const categoryItems = new Map<CategoryUnion, string>([
   ['ETC', '기타'],
 ]);
 
-export const levelItems = new Map<LevelUnion, string>([
+export const levelItems = new Map<Level, string>([
   ['BEGINNER', '입문'],
   ['LOW', '초급'],
   ['MIDDLE', '중급 이상'],
