+++ src/containers/pages/explore/constants/filter.constant.ts
@@ -1,7 +1,8 @@
 import { categoryItems, levelItems } from '@/constants/type-label-map';
-import { Category, Level } from '../models';
+import { Category } from '@/apis/models/video';
+import { FilterLevel } from '../models';
 
-export const levelFilterItems = new Map<Level, string>([
+export const levelFilterItems = new Map<FilterLevel, string>([
   ['all', '전체'],
   ...Array.from(levelItems),
 ]);
